pyproject.toml
[tool.poetry]
name = "rf-platform"
version = "0.1.0"
description = "RF Parameter Analysis and Simulation Platform"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2"
djangorestframework = "^3.14"
celery = "^5.3"
redis = "^4.5"
numpy = "^1.24"
scipy = "^1.10"
pandas = "^2.0"
requests = "^2.31"
django-redis = "^5.3"
psycopg2-binary = "^2.9"
gunicorn = "^21.2"

[tool.poetry.dev-dependencies]
pytest = "^7.4"
pytest-django = "^4.5"
black = "^23.3"
isort = "^5.12"
flake8 = "^6.0"
mypy = "^1.3"
coverage = "^7.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 100

[tool.mypy]
python_version = "3.9"
strict = true
ignore_missing_imports = true
